Se refactorizan las funciones de plot_functions en una clase
